Improve finding Python library/includes
On macOS, `find_package(PythonLibs ...)` does not find the Python
include   directory or library when they were installed via homebrew.
`find_package(Python ...)` is more robust, but only available on newer
versions of CMake. If we are compiling cvc5 on macOS and we have a newer
version of CMake available, this tries to use `find_package(Python ...)`
to find the paths.