[Parser] Rename generated `.c` to `.cpp` files
Using Clang on my machine, I started getting the error:

```
error: invalid argument '-std=c++17' not allowed with 'C'
```

This seems to be because we are treating the generated `.c` files as C++
files and it seems to break under certain conditions. To make our build
system more robust, this commit changes our command for generating files
to rename the generated files to use the `.cpp` extension.