Fixes and refactoring options passed to subsolvers (#8972)
This fixes some issues related to options passed to subsolvers.

We currently have issues due to the fact that set defaults is run on the same options objects twice, leading to unexpected behavior. This patches the issue to avoid some of these bad behaviors, including unsat-cores and abducts being incompatible, which is corrected in this PR.

It makes certain calls to subsolvers more consistent.

Further PRs will correct the issue so that original options (pre set defaults) are passed to subsolvers.