Fix spurious registration lemma for string disequality split (#9014)
Fixes #9003.

This removes a lemma that said that one of our splitting skolems could be assumed to have length >= 1, introduced here: #4821.
This lemma is unsound, since it assumes the length of a skolem, which can be introduced for other reasons. In other words, the original form of the lemma is not valid. I'm disabling this lemma altogether, as I don't think its necessary for performance (we constrain the length of the skolem in the conclusion anyways).