[FP] Address `// Removed` comments (#9046)
Fixes cvc5/cvc5-wishues#61. The FP rewriter
removes certain kinds, such as fp.eq. The FP solver (rightfully) did
not declare these kinds as function kinds. However, the // Removed
comments for those kinds were a bit mysterious. This commit makes the
assumption expressed by those comments explicit: It asserts that the
kind of each node passed to ppRewrite() is to not one of the kinds
that we assume to have been removed.