insert system-property to override loading of native libraries.
Depending on the environment and application context,
the user might want to override the loading-process for native libraries
and apply his own strategy for finding and loading the required native library.
Examples for this are:
 - loading the library from a different application-specific path,
 - extracting it from ZIP file and then loading it.
In those cases, the user will load CVC5 in the application context,
and then CVC5-Solver tries the direct way and might crash.

The new property 'cvc5.skipLibraryLoad' simply deactivated the CVC5-internal loading.
The default behaviour (without setting the property) is a direct loading from system-/library-path.

The new property and its handling are based on a similar solution in Z3 that can be found here:
https://github.com/Z3Prover/z3/blob/a3161bdc155f66a619ae740887cb267e57dd0e97/scripts/update_api.py#L624

Signed-off-by: Karlheinz Friedberger <kfriedberger@gmail.com>