[FP] Do not attempt to word blast within closures (#9055)
Fixes cvc5/cvc5-projects#477. Fixes
cvc5/cvc5-projects#509. It can happen that
certain closures such as set comprehension appear within a
floating-point term. Currently, the word blaster descends into the
closure and tries to word blast the body of the closure including bound
variables, which leads to the issues mentioned. This commit changes the
behavior to prevent the FP solver from descending into closures (both
for word blasting and computing models). Instead, we rely on the theory
responsible for the closure to reduce it in an appropriate manner.