Ensure that facts are rewritten
Currently, we do not ensure that facts are rewritten. This can lead to
terms being added to the equality engine in their unrewritten and
rewritten form:

```
...
  [(+ i (* (- 1) (str.len z_6)))]
    SKOLEM_FUN_STRINGS_DEQ_DIFF_56
    SKOLEM_FUN_STRINGS_DEQ_DIFF_16
    (+ i (* (- 1) (str.len y1)))
  [(- i (str.len z_6))]
...
```

This commit adds an assertion that facts must be rewritten. It also
modifies the inference manager for the theory of strings to rewrite an
inference before determining whether an inference can be treated as a
fact.