Add basic utilities for instantiation evaluation (#9011)
This is the initial step towards merging the instantiation evaluation branch: https://github.com/ajreynol/CVC4/tree/instEval

This branch introduces new techniques for incrementally maintaining information regarding the status of instantiations (for example, whether they are already entailed in the current context). This is an improvement over the current tests, which are done only after instantiations are constructed. This will allow (any) instantiation strategy to backtrack at the moment an instantiation is redundant, which in practice can significantly impact the time it takes to construct all instantiations that it should consider.

This PR introduces the basic utilities towards this branch. Most of the implementation of these utilities is omitted for now.