Move arithmetic-bitvector conversions to the UF theory (#9096)
This moves the non-standard kinds BITVECTOR_TO_NAT and INTEGER_TO_BITVECTOR to the UF theory.

The justification for this move is that neither the BV or arithmetic theories should have to deal with function symbols that have children that belong to another theory. Moreover they do not have the "parametric" property in their kinds file. This makes theory combination more complicated when these symbols are present in assertions.

Currently, they are eliminated eagerly during preprocess, which sidesteps issues of theory combination. The move to UF will allow lazy approaches to handling these functions, which will be done as a new subsolver of the UF theory.

This is the first step towards merging this branch: https://github.com/ajreynol/CVC4/tree/arithBvConvUf
This branch is
+22-0 on the benchmarks from #8786 (gains +22 on 59 benchmarks)
+34-22 on the benchmarks from #9052
Solves the benchmark on #8958 instantly.

Note this implies UF must be used in logic strings when these functions are used.

FYI @barrettcw