Parse S-expressions as terms
Fixes #8900. Currently,

```
(set-option :diagnostic-output-channel "stdout")
```

creates a `"stdout"` file, because the parser parses all S-expressions
(except nested ones) as strings. However, for values used in
`set-option` commands, this causes a problem because string values are
stored including the double quotes, which throws off our option handlers
that do not expect those double quotes. However, we cannot simply remove
the double quotes for strings in S-expressions, because otherwise some
S-expressions are printed incorrectly (e.g., `(set-info :foo "bar bar")`
would be printed as `(set-info :foo bar bar)`, which is incorrect).

This commit implements a solution that refactors the parser to parse
S-expressions as terms. For the `set-option` command, the commit removes
the double quotes around strings. As a result, `-o raw-benchmark` may
dump `set-option` commands with the double quotes removed, which is not
perfect, but could be fixed by further refactoring `SetOptionCommand` to
store a `Term` instead of an `std::string` for the value.

This change allows us to share some code for parsing "special constants"
in terms and S-expressions as defined in the SMT-LIB standard. The
commit also fixes a minor issue where whitespace characters were
expected to be escaped, which is incorrect according to the SMT-LIB
standard (the previous code called an ad-hoc string escape routing that
handled these cases before the string was stored in a
`cvc5::internal::String`). Finally, the change requires us to be careful
when generating strings from S-expressions: real values must be printed
as decimal numbers and not fractions.