Remove expandDefinitions infrastructure in the preprocessor (#9034)
Users of the method Preprocess::expandDefinitions included several more advanced features (e.g. abduction, interpolation) which should not provide input to subsolvers that involves internally introduced symbols (which ExpandDefinitions::expandDefinitions does). This corrects the issue.

This simplifies the functionality of the preprocessor to apply substitutions only.

Expand definitions is still required for getValue, where it is now used as a local utility.