Add SMT driver class (#9082)
This introduces the SMT driver class, which is layer between SolverEngine and SmtSolver. It will be used for satisfiability algorithms that may require multiple check-sat calls, or other interactions between SmtSolver and e.g. the preprocessor.

Followup PRs will connect it to SolverEngine and refactor the code for deep restarts to fit into this framework.