Reflect function definitions in LFSC. (#9109)
This PR modifies the LFSC proofs to print non-recursive function definitions as LFSC definitions instead of declarations and assumptions. The definitions are passed to the LFSC postprocessor and printer in a type-safe manner by adding a global scope in the proof node introducing the definitions. Since LFSC definitions are just macros, the LFSC postprocessor replaces applications of the ASSUME rule for definitions with REFL.