[CMake] Use `SYSTEM` includes for dependencies (#9106)
This changes our includes of dependencies to be `SYSTEM` includes,
which disables warnings originating from those includes. The current
version of CoCoA leads to warnings that cause issues when
compiler warnings are treated as errors.

This commit also enables testing of the `--cocoa` option in CI to ensure
that such issues are discovered earlier.