Fix get-value output to not letify across key-value pairs (#9127)
The ioutil properties are currently being ignored when printing cvc5::Terms. This leads to our response to get-value letifying across key/value pairs, which is wrong.

This corrects the issue by printing key values independently.

It also removes the ioutil calls from command.cpp (which are not effective currently). This means that we are now (intentionally) using the dag-thresh setting of the options to determine how terms in these commands are printed.

We should consider fixing operator<<(ostream, Term) to carry the properties of the given output stream, which will be done in a followup PR if necessary.