[Arith] Support integer in inverse functions (#9134)
Fixes #9131.

We eliminate inverse functions such as arcsin(x) by replacing them with a fresh skolem sk and adding a range lemma and an equality x = sin(sk). However, x may be an integer term, so we have to be careful about types when constructing the equality. This commit fixes the issue by using the existing helper method mkEquality().

Additionally, the internal type checking example led to an assertion failure decisionLevel() == 0, because the decision trail was not reset after the type exception was thrown. This commit makes the code in SmtDriver more robust to avoid such secondary assertion failures.