Fix computation of cardinality class
Fixes https://github.com/cvc5/cvc5-projects/issues/435. To compute the
cardinality class of constructors, we compute the maximum of the
cardinality of the arguments. However, we considered the result of
`max(INTERPRETED_ONE, FINITE)` to be `FINITE`. This is not accurate when
finite model finding is turned off. This led to issues in the sequences
solver, because we considered a constructor with an argument of an
uninterpreted sort as finite and tried to retrieve that cardinality.
This commit fixes the issue by making the result of
`max(INTERPRETED_ONE, FINITE)` be `FINITE_INTERPRETED`.

Note in the original issue, the following assertion failure:

```
Fatal failure within void cvc5::Minisat::Solver::pop() at src/prop/minisat/core/Solver.cc:1971
Check failure

 decisionLevel() == 0
```

occurred because of an `IllegalArgumentException` was being thrown
(calling `Cardinality::getFiniteCardinality()` on a sort with an
non-finite cardinality). This exception then lead to an unexpected
decision level (similar to the issue with uncaught
`TypeCheckingExceptionPrivate` exceptions fixed in
c4be9f55771b4babbaddf60f9ff215695562efde). There is an on-going effort
to get rid of `IllegalArgumentException`, so it is not worth adding a
fix for this. For example, commit
f8306b27c11abf4f758fbe78e40481494e05ee22 changed the code in
`cardinality.cpp` that was throwing the exception in this case.