[proof] [sat] Avoid duplicating assumption in chain links for SAT proof (#9141)
Previously the generation of proofs in the SAT proof manager was generating different ASSUMPTION proof nodes for the same assumptions. This led to duplication in the final SAT proof of the proofs of these respective nodes. When these nodes are justified via something whose post-processing is expensive, such as MACRO_RESOLUTION, we had performance issues.

This commit addresses the issue by using a specialized buffered proof generator for the links that when chained compose the SAT proof. It caches and reuses ASSUMPTION proof nodes when generating the proof nodes for the respective conclusion of the links. In our regressions this change leads to a 17% reduction in total size of proofs.

This commit also fixes an issue in the ProofCnfStream where we were adding symmetry steps automatically, which could be problematic, as explained in the added documentation. This issue happens in our regression regress1/strings/issue6184-unsat-core.smt2 after the changes for size reduction, and is (thus indirectly) duplicated by cvc5/cvc5-projects#529. Likewise, we were not properly accounting for disabled symmetry in the LazyCDProof when asking for a generator for a fact. This commit also fixes this.