[parser] Fix handling of the :named attribute
We must bind the name and the named term in the parser symbol table at the
moment the :named attribute is processed, otherwise we will not be able to parse
problems which use the name within the same command (which is allowed by the
SMT-LIB semantics).

This commit restores the behavior (changed in #9025) by allowing
DefineFunctionCommand (via which we track the :named attributes) to be
configurable to not necessarily bind the defined function and its definition in
the symbol table.