Fix getEqualityStatus for non-linear arithmetic terms (#9148)
Fixes the 1st and 3rd benchmarks on #9113.

Previously we were using substitutions that involve non-linear terms, e.g. applications of iand would assume the value assigned to them by the linear arithmetic solver.

This changes a check when computing the model substitution to check whether the domain of the substitution is leaf. Previously, we were checking the range, likely by mistake.