NM: Initialize NodeManager on construction. (#9154)
The NodeManager is a true singleton and thus only a single instance ever exists. Thus it should be initialized on construction.
This removes an assertion in NodeValue::inc(), which implicitly calls NodeManager::currentNM() and thus causes infinite recursion when the NodeManager is initialized on construction due to the insertion of the null node into the node value pool. @ajreynol and I suspect that this assertion was for previous incarnations of multi-threading support, which makes this assertion obsolete.